producers/requirements.txt
pulsar-client
pyhive
flask
websocket-client
python-dotenv
pandas
schedule